Rorg main method and add basic usage to README.
