+++ README.md
@@ -1,3 +1,9 @@
 Pipelined FFT [![Build Status](https://travis-ci.org/ucb-art/fft.svg?branch=master)](https://travis-ci.org/ucb-art/fft)
 =======================
 
+# Usage
+
+Run `sbt test` to run the tests in `src/test/scala`.
+
+Run `sbt run` to run the code in Main.scala. This runs the FFT generator and saves the verilog output to `generated-src`.
+
